New package and repo location
Moving the location of the project to be under the sosreport
organization on GitHub. Additionally, renaming the project from
'clustersos' to 'sos-collector' to signify that this tool is not tied to
any specific clustering technology.

sos-collector is designed to collect sosreports from multiple nodes in
an environment. These environments often employ some form of
'clustering', so when an end user needs to collect sosreports they often
need to collect one from each (or at least several) node.

sos-collector works to eliminate the need for manual runs of sosreport
multiple times, by enumerating the nodes in the cluster and then running
sosreport on each node and collecting the resulting archives in a single
tarball where sos-collector was run. This means that end users and
support representative need only one command and get a single collection
of sosreports rather than dealing with multiple files.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>