[sos-collector] Improve handling of sos plugin options
Improves the logic behind formatting the sosreport command that gets
executed on nodes by enabling the --only-plugins option properly, and
comparing the commandline --skip-options to --enable-options and
removing any skipped plugins. This allows the user to override cluster
profiles for enabled plugins.

These sosreport options can also now be specified multiple times on the
commandline. E.G. both

	sos-collector -n logs,processor
and
	sos-collector -n logs -n processor

are valid and will result in the same behavior.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>