[sos-collector] Extend sos option coverage
Adds additional sosreport options that sos-collector can take and pass
to the sosreport command executed on nodes. This includes the chroot,
sysroot, verify, and compression options.

Further, adds a 'sos-cmd' option that overrides all sosreport command
formatting that sos-collector does, aside from the usage of --batch, to
allow a user to manually define exactly what will be run on the nodes.

With regards to #3, we can validate options like chroot and compression
via argparse, whereas the validation of skipped/enabled plugins should
be handled by sosreport, with the assumption that the user knows valid
plugins to specify. For 'sos-cmd' we accept that if this option is used,
the user specifying the command knows what they are doing and we do not
do any validation.

Closes: #3

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>