[sosnode] Correct determination of final sosreport path
There are some conditions where the sosreport name determined by
sos-collector could previously include a newline (or potentially other)
characters which broke retrieving the sosreports from remote nodes.

This adds a catch-all to only take the first part of a file path as the
location of the sosreport to avoid this situation.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>