[sosnode] Add sos version checks and command building validation
Adds basic sos command validation on a per-node basis. First, sos is
verified to actually be on the node, and from there use --list-plugins
output to get a list of plugins (enabled and disabled) and available
plugin options for the enabled plugins.

Now, each node validates the requested options to enable or disable
plugins, and the setting of plugin options. If the plugin does not
exist on the node, it is stripped from the final sos command that is run
rather than being blindly applied which would cause sos to immediately
exit when a non-existing plugin was specified on the node.

We further try to validate that the plugin options requested are
available as well.

This will allow for less sosreport failures due to differing sos
versions on nodes.

Closes: #6

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>