[ovirt] Collect database by default, add no-hypervisors option
Makes collection of the database automatic by default, to mimic
log-collector which does not prompt the user for the db information. The
assumption here is that if the user has administrative rights on the
manager, they should be able to parse the configuration to get the db
credentials.

Also adds a 'no-hypervisors' option to skip node enumeration.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>