[sosnode] Fix Atomic Host execution
Fixes several bugs around running on Atomic Host systems.

First, allows the new sos validation to function properly by prefixing
the relevant commands with the proper 'atomic' commands. Second,
corrects an issue that could cause sos-collector to not properly parse
the returned file path for the sosreport archive.

As part of this correction, set_sos_prefix() has been moved to
SosNode(), instead of being set within the cluster profiles. This makes
more sense as it is more likely that this will need to be controlled by
host OS distribution than clustering technology.

Finally, this change also sets the container image to be used to be
controlled by the determined distribution of the node.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>