[sos_collector] Prevent duplicate or conflicting plugin options
Changes the way plugin options from clusters are handled to prevent them
from duplicating or conflicting with commandline args from the end user.
If a plugin option set by a cluster profile is also set on the
commandline, the cluster option is not added - instead the commandline
option is always used instead.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>