[sos-collector] Update name/label option to match sosreport
Updates sos-collector's name option to be consistent with sosreport's
label option, which replaced the name option. Additionally, improves the
way clusters can set node labels by also setting master labels if
appropriate.

Users are no longer prompted for their names during execution. If a user
wants to add a label to the sos-collector archive and the sosreports
generated on the nodes, the commandline --label option should be used.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>