[pacemaker] Additional offline node check and logic fixes
Adds another line check for offline nodes, for which pacemaker has
separate reports for "offline UNCLEAN" nodes and "OFFLINE" nodes, where
the latter means that the cluster is not running on the node even if it
is reachable.

The pacemaker profile now also immediately returns an empty node list if
the given "master" node is not able to report the cluster status (e.g.
the cluster is not running on that node).

Additionally, fixes the logic in determining these nodes to reflect what
the cluster profile options are set to.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>