[sosnode|ovirt] Correct ovirt node labels
The previous check for if a node was a 'regular' installation or a
hypervisor installation was valid only in the 3.x lifecycle of
ovirt/RHEV. Update this to reflect the current 4.x version.

This required moving the is_installed() method from the Cluster() to the
individual SosNode() instantiations. Realistically, this should have
always been the case. Along with this move, cluster enablement has been
update to explicitly call only the master node's is_installed() method
for package checks.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>