[sosnode] Fix execution when connecting as non-root user
Fixes execution of commands run on remote nodes when those commands need
root privileges and we're not connecting as root. Adds a new parameter
to the run_command() and exec_master_command() methods called
'need_root' which will handle properly formatting the command to use
either 'su -c' or 'sudo' based on configuration, and then provide that
password automatically over the SSH session.

This also updates the ovirt cluster profile to properly use the new
style of running commands as non-root.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>