[sos-collector] Add insecure sudo option
Adds an 'insecure-sudo' option for when passwordless sudo is configured
on the nodes. This will skip the sudo password prompt and further
protect against erroneously sending input to a remote command execution
when using sudo.

Additionally, moves the root password prompt when using --become to
after the SSH prompt to provide a more consistent experience.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>