[clusters] Allow for layered products as distinct clusters
Cluster profiles can now contain multiple class definitions that can be
used for layered products building off of lower-level technologies.

For instance, the kubernetes profile now also contains an openshift
cluster definition since openshift builds on kubernetes and all we need
to do is define a new set of packages to activate on and a new command
to use for node enumeration.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>