[ovirt] Rework database query
The database query used to get a node listing has been reworked to be
able to accept both the cluster and datacenter options simultaneously.
In practice, it would be extremely rare to need both, but the
possibility exists and it also makes further maintenance of the database
query much simpler.

Default behavior remains to select all hypervisors in the environment.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>