[security] Handle attempts at shell code in options
With the --sos-cmd option, users can pass in a string to be used as a
sosreport command. This is useful for several use cases but previously
went without validation, partly due to the complexity of validating the
raw string against sos options.

Of bigger concern however is that a malicious user could inject shell
code to this option, e.g. "--sos-cmd='--bogusoption ||
malicious_code_here'", which is a big concern for sos-collector.

Now, we scan the provided sos command string for telltall shell script
characters, and if they are detected we reject the entire option value.
This is because it is not feasible to validate a string once a positive
identification is made. For now, sos-collector will run normally after
discarding the sos-cmd option, and the user will be alerted to this
behavior. We may end up changing this behavior in the future and end up
aborting execution after making a positive detection for potential shell
script.

Additionally, it is theorectically possible for this same attack vector
to be used by cluster option values, though this was not seen to be
useable with the current set of options. Nonetheless, we now will drop
anything after a space character in a cluster option as no cluster
option currently has any legitimate use for such values.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>