[sosnode] Don't try to open an sftp session locally
With a previous commit, we changed how we open/manage sftp sessions over
paramiko. This introduced an issue where we tried to do this even when
a node was the local host. Now we do not try to do this.

Additionally, correct the logic for handling cluster setup and extra
command gathering in the specific instance where all the following are
true - sos-collector is run on a local node without the master option,
this local node is not able to enumerate the cluster nodes, and the node
list and cluster type are provided by the user.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>