[sosnode] Improve file handling methodology
Standardizes the way we handle collecting and removing files from nodes.

Additionally, collects the md5 sums of any generated sosreports for the
final archive.

Finally, adjusts the way we handle collecting extra information from a
cluster outside of the per-node sosreports. Combined this should provide
better stability for generation and collection of extra information, for
example the ovirt database dump.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>