[sos_collector] Configure initial sosreport command after case id prompt
We were previously configuring the initial sosreport command before the
case id prompt when a user didn't specify either --case-id or --batch on
the commandline, thus only the sos-collector archive would have the case
id and not the sosreports run on the nodes. No other options were
affected by this, and case ids are now properly set on all node
sosreports.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>