[build] Re-add license to setuptools config
The previous change to remove the license from setup.py was incorrect as
it turns out that a previous version of setuptools was erroneously
adding the license even without it being specified in setup.py. Newer
versions of setuptools correct this behavior, so we need to add the
license back in order to get successful rpm builds.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>