[sos_collector] Do not attempt to delete temp log files that get moved
On certain hosts, sos-collector could generate unexpected exceptions at
the end of a successful run that indicated an attempt to remove the
temporary files created for sos-collector's own logging. These temp
files would have already been moved to the final archive, and thus the
attempts to delete ended in a FileNotFoundError.

Since the files were moved, and the temp directory as a whole is removed
during cleanup, we can simply avoid the automatic attempt to delete the
file and avoid the exceptions.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>