[soscollector] Fix race condition in loading cluster profiles
Fixes a race condition where a layered cluster profile would not be
properly loaded because the base profile would be checked first.

Now, once a profile is matched, we check to see if there are any layered
profiles built on top of it. If there are, those are checked and if
their enablement check returns True, the layered profile is used instead
of the base profile. If the layered profile's check returns False, or if
no layered profiles are found, the base profile is still used.

Resolves: #12

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>