[global] Quote options locally everywhere
An earlier commit passed cluster option values through pipes.quote() in
an effort to prevent command injection on the remote nodes, however this
did not capture every avenue through which a maliciously designed option
value or sos command could take.

Now, quote the option values at the time of use, and also quote any sos
option that isn't an on/off toggle.

Additionally, re-work how the ovirt/rhv database query is quoted and
filter out cluster/datacenter values that are likely to contain SQL
code, since we cannot rely on the driver to do this since we don't have
an actual connection to the database.

Original discovery of issues and patch work thanks to Riccardo Schirone.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>