[global] Use SSH ControlPersist instead of Paramiko
Paramiko has existing security concerns, and in addition to that a
history of security concerns. Among other issues, this means that FIPS
systems cannot include paramiko or any tools that incorporate the
library.

This change moves sos-collector away from paramiko and instead directly
leveraging SSH ControlPersist functionality, which was introduced into
OpenSSH in version 5.6, back in 2010. At this point, it is reasonable to
assume that any supported host installation has at least that version
installed and supports ControlPersist.

This means that any host that does not have a system SSH installation
that supported ControlPersist is not able to run sos-collector. Sorry,
those are the breaks as in order to properly resolve the ongoing
security concerns with paramiko, sos-collector cannot even have the
option of using that library.

However, this does give us a number of smaller benefits, including:

    - Time to open an SSH session is reduced
    - Overall performance of running remote commands should improve
    - Any SSH keytype supported by the host installation is supported

On the topic of SSH keys, it should be noted that Paramiko does not
support the OPENSSH PRIVATE KEY syntax for RSA or DSA keys, which as of
the most recent version of OpenSSH is now the default style of SSH key
generated. By using the system installation, this is not a concern for
us any longer.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>