[satellite] Add satellite a supported environment
Adds a profile for Satellite 6. While not technically a cluster, Sat 6
does have capsule servers that can be deployed concurrently to help with
load and distribution, so it still fits in with sos-collector's goal.

Many thanks to Paul Dudley from Red Hat for assistance in providing a
test environment for this work and identifying the need.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>