[sosnode] Improve container command execution
Updates the container execution flow to reuse a temporary container that
we start when the host is first determined to be a container OS.

From there, the container is kept running and subsequent commands are
run in that container's namespace if needed.

This allow us more flexibility when adding containerized hosts in the
future, as well as reducing the cost associated with running sos-related
commands inside containers on those types of hosts, as previously we had
to effectively re-launch the container for every command we were
running.

This Atomic and RHCOS host profiles have been updated to use this new
approach, and their cleanup operations now use --force to make sure the
SPC is killed when we're finished, regardless of what may be occurring
inside that container (e.g. sosreport still running when we hit a
timeout).

Resolves: #18

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>