[clusters] Add a 'no-cluster' cluster profile
Adds a profile to allow users to specify "none" or "jbon" (just of a
bunch of nodes) via --cluster-type to avoid any and all cluster-specific
checks or enumeration.

This can be used when providing a list of nodes that are not clustered
together in any meaningful or deterministic way, and still allow
sos-collector to collect from those nodes without having to specify a
bogus cluster-type which could affect how sosreport is run on the nodes.

When using this cluster-type, _only_ the nodes provided by the --nodes
option will be used. Note that this means regex string(s) provided to
the --nodes option will not be evaluated and will effectively be
ignored.

Resolves: #19

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>