[sos-collector] Ability to use 'host group' files
This change adds the ability to use and write JSON-formatted files to
use as 'host groups' that define cluster-type, master, and a list of
nodes to use for collection.

Using the new --group option, sos-collector will look to
/var/lib/sos-collector for a host group with a matching name to load and
use for execution. Note that users may still use the --nodes option when
using --group - the two lists of nodes will be merged.

Using the new --save-group option will write out a host group file to
/var/lib/sos-collector/ with the same name as the value given to this
option. The values written to the host group file will be what was
determined for this run of sos-collector; meaning the cluster-type,
master node, and node list enumerated from the cluster. This means that
regex strings will not be saved to group files, but the results of those
regexes will be saved.

Fixes: #20

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>