[ovirt] Add RHHI-V support
Adds support for RHHI-V environments which are RHV environments that
also use the hypervisors as gluster nodes. The new 'rhhi_virt' cluster
profile will be enabled when sos-collector is run against an environment
that is _both_ RHV and has nodes listed in the 'gluster_server' table in
the RHV database. Note that this means if community oVirt is in use, the
gluster bits enabled by 'rhhi_virt' will not be enabled for the oVirt
cluster type.

Included with this change is making DB queries more programmatic, and
making minor stylistic changes to the main query used by get_nodes() to
allow easier reading of the SQL query being built.

Finally, remove an unused import of getpass.

Resolves: #21

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>