[ovirt] Use engine ssh key if running locally on manager
If running on an oVirt/RHV manager, the SSH keys that the manager uses
to connect to the hypervisors are available to us. Like
ovirt-log-collector, sos-collector should leverage these so that users
don't need to manually specify them, distribute other keys, or provide
the password for the nodes.

Now if the ovirt cluster (or a layered profile) is loaded, *and*
sos-collector is being run locally on the manager, *and* the user does
not provide either an SSH key to use or one of the password options,
sos-collector will use the engine key at
/etc/pki/ovirt-engine/keys/engine_id_rsa.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>