[ovirt] Add spm-only option
Adds an option to the ovirt profile to only collect from SPM nodes,
which there will be one for every datacenter in the environment.

Included with this change is referencing the 'vds' table in the database
instead of 'vds_static' as the former includes a per-node state of the
SPM role where as the latter does not.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>